requirements.txt
accelerate
causal-conv1d
datasets
liger-kernel
pyyaml
transformers>=5.0.0
torch>=2.6.0
wandb
